moving some test files around
