+++ src/Tests.ToolKit/Data/Mongo/DataRepositorySpecs/CreateItemTests.cs
@@ -3,7 +3,7 @@ using FatCat.Toolkit.Testing;
 using FluentAssertions;
 using Xunit;
 
-namespace Tests.FatCat.Toolkit.Data.DataRepositorySpecs;
+namespace Tests.FatCat.Toolkit.Data.Mongo.DataRepositorySpecs;
 
 public class CreateItemTests : DataRepositoryTests
 {
