+++ src/Tests.ToolKit/Data/Mongo/DataRepositorySpecs/DeleteItemTests.cs
@@ -4,7 +4,7 @@ using FluentAssertions;
 using MongoDB.Driver;
 using Xunit;
 
-namespace Tests.FatCat.Toolkit.Data.DataRepositorySpecs;
+namespace Tests.FatCat.Toolkit.Data.Mongo.DataRepositorySpecs;
 
 public class DeleteItemTests : DataRepositoryTests
 {
