+++ src/Tests.ToolKit/Data/Mongo/DataRepositorySpecs/GetFirstOrCreateTests.cs
@@ -5,7 +5,7 @@ using FluentAssertions;
 using MongoDB.Driver;
 using Xunit;
 
-namespace Tests.FatCat.Toolkit.Data.DataRepositorySpecs;
+namespace Tests.FatCat.Toolkit.Data.Mongo.DataRepositorySpecs;
 
 public class GetFirstOrCreateTests : DataRepositoryTests
 {
